fix(java): align version type from i32 to u64 (#5892)
This should be a mistake introduced long long ago

+++ java/src/main/java/org/lance/Dataset.java
@@ -341,7 +341,7 @@ public class Dataset implements Closeable {
 
   private static native Dataset openNative(
       String path,
-      Optional<Integer> version,
+      Optional<Long> version,
       Optional<Integer> blockSize,
       long indexCacheSize,
       long metadataCacheSizeBytes,
