+++ java/src/main/java/org/lance/ReadOptions.java
@@ -25,7 +25,7 @@ import java.util.Optional;
 /** Read options for reading from a dataset. */
 public class ReadOptions {
 
-  private final Optional<Integer> version;
+  private final Optional<Long> version;
   private final Optional<Integer> blockSize;
   private final long indexCacheSizeBytes;
   private final long metadataCacheSizeBytes;
@@ -43,7 +43,7 @@ public class ReadOptions {
     this.storageOptionsProvider = builder.storageOptionsProvider;
   }
 
-  public Optional<Integer> getVersion() {
+  public Optional<Long> getVersion() {
     return version;
   }
 
@@ -87,7 +87,7 @@ public class ReadOptions {
 
   public static class Builder {
 
-    private Optional<Integer> version = Optional.empty();
+    private Optional<Long> version = Optional.empty();
     private Optional<Integer> blockSize = Optional.empty();
     private long indexCacheSizeBytes = 6 * 1024 * 1024 * 1024; // Default to 6 GiB like Rust
     private long metadataCacheSizeBytes = 1024 * 1024 * 1024; // Default to 1 GiB like Rust
@@ -101,7 +101,7 @@ public class ReadOptions {
      * @param version the version of the dataset
      * @return this builder
      */
-    public Builder setVersion(int version) {
+    public Builder setVersion(long version) {
       this.version = Optional.of(version);
       return this;
     }
